Move `load_checkpoint` to utils
