+++ python/ray/train/huggingface/__init__.py
@@ -1,6 +1,8 @@
 from ray.train.huggingface.huggingface_predictor import HuggingFacePredictor
 from ray.train.huggingface.huggingface_trainer import (
     HuggingFaceTrainer,
+)
+from ray.train.huggingface.utils import (
     load_checkpoint,
 )
