+++ python/ray/train/lightgbm/__init__.py
@@ -1,6 +1,6 @@
 from ray.train.lightgbm.lightgbm_predictor import LightGBMPredictor
-from ray.train.lightgbm.lightgbm_trainer import LightGBMTrainer, load_checkpoint
-from ray.train.lightgbm.utils import to_air_checkpoint
+from ray.train.lightgbm.lightgbm_trainer import LightGBMTrainer
+from ray.train.lightgbm.utils import to_air_checkpoint, load_checkpoint
 
 __all__ = [
     "LightGBMPredictor",
