+++ python/ray/train/rl/__init__.py
@@ -1,4 +1,5 @@
 from ray.train.rl.rl_predictor import RLPredictor
-from ray.train.rl.rl_trainer import RLTrainer, load_checkpoint
+from ray.train.rl.rl_trainer import RLTrainer
+from ray.train.rl.utils import load_checkpoint
 
 __all__ = ["RLPredictor", "RLTrainer", "load_checkpoint"]
