+++ python/ray/train/sklearn/__init__.py
@@ -1,5 +1,5 @@
 from ray.train.sklearn.sklearn_predictor import SklearnPredictor
-from ray.train.sklearn.sklearn_trainer import SklearnTrainer, load_checkpoint
-from ray.train.sklearn.utils import to_air_checkpoint
+from ray.train.sklearn.sklearn_trainer import SklearnTrainer
+from ray.train.sklearn.utils import to_air_checkpoint, load_checkpoint
 
 __all__ = ["SklearnPredictor", "SklearnTrainer", "load_checkpoint", "to_air_checkpoint"]
