+++ python/ray/train/sklearn/sklearn_predictor.py
@@ -8,7 +8,7 @@ from ray.air.checkpoint import Checkpoint
 from ray.air.constants import TENSOR_COLUMN_NAME
 from ray.train.predictor import Predictor
 from ray.train.sklearn._sklearn_utils import _set_cpu_params
-from ray.train.sklearn.sklearn_trainer import load_checkpoint
+from ray.train.sklearn.utils import load_checkpoint
 from ray.util.joblib import register_ray
 
 if TYPE_CHECKING:
