+++ python/ray/train/tensorflow/tensorflow_trainer.py
@@ -1,9 +1,8 @@
-from typing import Callable, Optional, Dict, Tuple, Type, Union, TYPE_CHECKING
-import tensorflow as tf
+from typing import Callable, Optional, Dict, Union, TYPE_CHECKING
 
 from ray.train.tensorflow.config import TensorflowConfig
 from ray.train.trainer import GenDataset
-from ray.train.data_parallel_trainer import DataParallelTrainer, _load_checkpoint
+from ray.train.data_parallel_trainer import DataParallelTrainer
 from ray.air.config import ScalingConfig, RunConfig, DatasetConfig
 from ray.air.checkpoint import Checkpoint
 from ray.util import PublicAPI
@@ -185,27 +184,3 @@ class TensorflowTrainer(DataParallelTrainer):
             preprocessor=preprocessor,
             resume_from_checkpoint=resume_from_checkpoint,
         )
-
-
-def load_checkpoint(
-    checkpoint: Checkpoint,
-    model: Union[Callable[[], tf.keras.Model], Type[tf.keras.Model], tf.keras.Model],
-) -> Tuple[tf.keras.Model, Optional["Preprocessor"]]:
-    """Load a Checkpoint from ``TensorflowTrainer``.
-
-    Args:
-        checkpoint: The checkpoint to load the model and
-            preprocessor from. It is expected to be from the result of a
-            ``TensorflowTrainer`` run.
-        model: A callable that returns a TensorFlow Keras model
-            to use, or an instantiated model.
-            Model weights will be loaded from the checkpoint.
-
-    Returns:
-        The model with set weights and AIR preprocessor contained within.
-    """
-    model_weights, preprocessor = _load_checkpoint(checkpoint, "TensorflowTrainer")
-    if isinstance(model, type) or callable(model):
-        model = model()
-    model.set_weights(model_weights)
-    return model, preprocessor
