+++ python/ray/train/tests/test_huggingface_trainer.py
@@ -14,7 +14,7 @@ from transformers.trainer_callback import TrainerState
 import ray.data
 from ray.train.batch_predictor import BatchPredictor
 from ray.train.huggingface import HuggingFacePredictor, HuggingFaceTrainer
-from ray.train.huggingface.huggingface_utils import TrainReportCallback
+from ray.train.huggingface._huggingface_utils import TrainReportCallback
 from ray.train.tests._huggingface_data import train_data, validation_data
 
 # 16 first rows of tokenized wikitext-2-raw-v1 training & validation
