+++ python/ray/train/torch/__init__.py
@@ -9,7 +9,7 @@ except ModuleNotFoundError:
 
 from ray.train.torch.config import TorchConfig
 from ray.train.torch.torch_predictor import TorchPredictor
-from ray.train.torch.torch_trainer import TorchTrainer, load_checkpoint
+from ray.train.torch.torch_trainer import TorchTrainer
 from ray.train.torch.train_loop_utils import (
     TorchWorkerProfiler,
     accelerate,
@@ -20,7 +20,7 @@ from ray.train.torch.train_loop_utils import (
     prepare_model,
     prepare_optimizer,
 )
-from ray.train.torch.utils import to_air_checkpoint
+from ray.train.torch.utils import to_air_checkpoint, load_checkpoint
 
 __all__ = [
     "TorchTrainer",
