+++ python/ray/train/torch/torch_trainer.py
@@ -1,11 +1,8 @@
-from typing import TYPE_CHECKING, Callable, Dict, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Callable, Dict, Optional, Union
 
-import torch
-
-from ray.air._internal.torch_utils import load_torch_model
 from ray.air.checkpoint import Checkpoint
 from ray.air.config import DatasetConfig, RunConfig, ScalingConfig
-from ray.train.data_parallel_trainer import DataParallelTrainer, _load_checkpoint
+from ray.train.data_parallel_trainer import DataParallelTrainer
 from ray.train.torch.config import TorchConfig
 from ray.train.trainer import GenDataset
 from ray.util import PublicAPI
@@ -196,24 +193,3 @@ class TorchTrainer(DataParallelTrainer):
             preprocessor=preprocessor,
             resume_from_checkpoint=resume_from_checkpoint,
         )
-
-
-def load_checkpoint(
-    checkpoint: Checkpoint, model: Optional[torch.nn.Module] = None
-) -> Tuple[torch.nn.Module, Optional["Preprocessor"]]:
-    """Load a Checkpoint from ``TorchTrainer``.
-
-    Args:
-        checkpoint: The checkpoint to load the model and
-            preprocessor from. It is expected to be from the result of a
-            ``TorchTrainer`` run.
-        model: If the checkpoint contains a model state dict, and not
-            the model itself, then the state dict will be loaded to this
-            ``model``.
-
-    Returns:
-        The model with set weights and AIR preprocessor contained within.
-    """
-    saved_model, preprocessor = _load_checkpoint(checkpoint, "TorchTrainer")
-    model = load_torch_model(saved_model=saved_model, model_definition=model)
-    return model, preprocessor
