+++ python/ray/train/xgboost/__init__.py
@@ -1,5 +1,5 @@
-from ray.train.xgboost.utils import to_air_checkpoint
+from ray.train.xgboost.utils import load_checkpoint, to_air_checkpoint
 from ray.train.xgboost.xgboost_predictor import XGBoostPredictor
-from ray.train.xgboost.xgboost_trainer import XGBoostTrainer, load_checkpoint
+from ray.train.xgboost.xgboost_trainer import XGBoostTrainer
 
 __all__ = ["XGBoostPredictor", "XGBoostTrainer", "load_checkpoint", "to_air_checkpoint"]
