+++ python/ray/train/xgboost/utils.py
@@ -1,9 +1,12 @@
 import os
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Tuple
 
 import xgboost
 
-from ray.air._internal.checkpointing import save_preprocessor_to_dir
+from ray.air._internal.checkpointing import (
+    save_preprocessor_to_dir,
+    load_preprocessor_from_dir,
+)
 from ray.air.checkpoint import Checkpoint
 from ray.air.constants import MODEL_KEY
 
@@ -34,3 +37,24 @@ def to_air_checkpoint(
     checkpoint = Checkpoint.from_directory(path)
 
     return checkpoint
+
+
+def load_checkpoint(
+    checkpoint: Checkpoint,
+) -> Tuple[xgboost.Booster, Optional["Preprocessor"]]:
+    """Load a Checkpoint from ``XGBoostTrainer``.
+
+    Args:
+        checkpoint: The checkpoint to load the model and
+            preprocessor from. It is expected to be from the result of a
+            ``XGBoostTrainer`` run.
+
+    Returns:
+        The model and AIR preprocessor contained within.
+    """
+    with checkpoint.as_directory() as checkpoint_path:
+        xgb_model = xgboost.Booster()
+        xgb_model.load_model(os.path.join(checkpoint_path, MODEL_KEY))
+        preprocessor = load_preprocessor_from_dir(checkpoint_path)
+
+    return xgb_model, preprocessor
